font manager에 대한 커밋
폰트 매니저 구현 종료.
문자들이 그려진 텍스처 리소스와 각문자의 uv, 크기, 빈공간의 간격등의 정보를 갖는 데이터파일이 존재.

이미지상의 대문자와 소문자의 간격 그리고 같은 라인안에 글자를 놓기위해선 각자 다른 정보가 필요하기 때문에 데이터 파일안에 문자별로 정보를 담아두고 있음.

fontShader에서 텍스트, 위치, 크기,간격정보를 갖는 구조체의 리스트를 받으면 정보에 맞게 텍스트를 그리는 구조.

하나의 버퍼에 각자 위치가 다른 문자열 데이터를 못넣기때문에(정보가 덮어써져서 마지막 데이터만 그리게 됨)
리소스뷰는 배열로 18개를 만들었다

즉, 프레임마다 그릴 수 있는 문자열은 최대 18개임(리스트의 각 문자열마다 하나의 뷰를 차지)

정점의 입력정보는 객체(문자)당 하나만을 갖고 정보에 있는 크기, vertexID를 통해 네개의 정점을 구하여 tri_strip으로 그려냄

+++ Client/DinoRun/CShaders/UiShader/GaugeShader.cpp
@@ -32,7 +32,7 @@ void GaugeShader::BuildObjects(shared_ptr<CreateManager> pCreateManager, void* p
 	m_ppObjects->SetMaterial(0, material);
 
 	PlaneMesh *mesh = NULL;
-	mesh = new PlaneMesh(0.003,0.06,0.1,0,1.0,0,1.0);
+	mesh = new PlaneMesh(0.003f,0.06f,0.1f,0.0f,1.0f,0.0f,1.0f);
 	mesh->CreateShaderVariables(pCreateManager->GetDevice().Get(), pCreateManager->GetCommandList().Get());
 
 	m_ppObjects->SetMesh(mesh);
@@ -54,7 +54,7 @@ void GaugeShader::BuildObjects(shared_ptr<CreateManager> pCreateManager, void* p
 void GaugeShader::Update(float fTimeElapsed, void* pInformation)
 {
 	CPlayer* pPlayer = (CPlayer*)pInformation;
-	UINT nGauge = pPlayer->GetMaxForce();
+	UINT nGauge = (UINT)pPlayer->GetMaxForce();
 	pPlayer->SetMaxForce(pPlayer->GetMaxForce() - fTimeElapsed * 2);
 	if (pPlayer->GetMaxForce() < MIN_FORCE) pPlayer->SetMaxForce(MIN_FORCE);
 	if (uvX[0] < (int)(nGauge * 0.05))
